shared/298: call fs commands on the loop device
In general calling fs tools is best done on the block device used for
the file system and not the backing device of a loop file.  Thus switch
shared/298 to call all fs commands on the loop device.  Also add a
common on why the xfs_io fiemap command is called on the backing file,
and to have a good place for the comment stop passing the backing file
as the argument to get_holes function and just use it implicitly as
the other helpers to with the loop device.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: "Darrick J. Wong" <djwong@kernel.org>
Signed-off-by: Zorro Lang <zlang@kernel.org>