xfs/558: scale blk IO size based on the filesystem blksz
This test fails for 64k filesystem block size on a 4k PAGE_SIZE
system. Scale the `blksz` based on the filesystem block size instead of
fixing it as 64k so that we do get some iomap invalidations while doing
concurrent writes.

Cap the blksz to be at least 64k to retain the same behaviour as before
for smaller filesystem blocksizes.

This fixes the "Expected to hear about writeback iomap invalidations?"
message for 64k filesystems.

Signed-off-by: Pankaj Raghav <p.raghav@samsung.com>
Tested-by: "Darrick J. Wong" <djwong@kernel.org>
Reviewed-by: "Darrick J. Wong" <djwong@kernel.org>
Signed-off-by: Zorro Lang <zlang@kernel.org>