fstests: guess proper MIN_FSSIZE when it's not set
This will fix test-zoned CI failures like:

  btrfs/156       [failed, exit status 1]1s - output mismatch (see /root/fstests/results//btrfs_zoned/btrfs/156.out.bad)
      --- tests/btrfs/156.out	2023-06-30 19:57:56.234014627 +0000
      +++ /root/fstests/results//btrfs_zoned/btrfs/156.out.bad	2024-05-15 01:01:09.649915920 +0000
      @@ -1,2 +1,7 @@
       QA output created by 156
      -Silence is golden
      +ERROR: size 1073741824 is too small to make a usable filesystem
      +ERROR: minimum size for a zoned btrfs filesystem is 47185920
      +mount: /mnt/scratch: wrong fs type, bad option, bad superblock on /dev/nvme1n2, missing codepage or helper program, or other error.
      +       dmesg(1) may have more information after failed mount system call.
      +mount /dev/nvme1n2 /mnt/scratch failed
      ...
      (Run 'diff -u /root/fstests/tests/btrfs/156.out /root/fstests/results//btrfs_zoned/btrfs/156.out.bad'  to see the entire diff)

Setting MIN_FSSIZE is necessary to skip a test, which create a FS with
a too small size. It is especially necessary for the zoned case. Without
that, there will be many failed tests.

Guess a proper MIN_FSSIZE if not set, and skip such tests.

Currently, it assumes that a FS will be formatted with "-d single -m dup"
(default) option. That can still convert all the RAID/DUP cases except
"-d dup -m dup".

Signed-off-by: Naohiro Aota <naohiro.aota@wdc.com>